Separate the page layout components
