+++ src/pages/ProjectsPage.tsx
@@ -16,9 +16,9 @@
 
 import React from 'react';
 import { Link } from 'react-router-dom';
-import { ProjectList } from 'features/projects';
+import { ProjectLists } from 'features/projects';
 
-export function Projects() {
+export function ProjectsPage() {
   return (
     <>
       <div className="border-b border-solid border-gray-200 mb-6">
@@ -43,7 +43,7 @@ export function Projects() {
           </Link>
         </div>
       </div>
-      <ProjectList />
+      <ProjectLists />
     </>
   );
 }
